[Refactor] group 타입파일 분리
